109 validation message

+++ src/app/Http/Controllers/SeekingController.php
@@ -54,13 +54,10 @@ class SeekingController extends Controller
 
     public function store(SeekingRequest $request)
     {
-        $imagePath = $request->file('seeking_thumbnail')->store('seeking_thumbnail', 'public');
-
         $seeking = new Seeking;
         $seeking->user_id = auth()->user()->id;
         $seeking->title = $request->title;
         $seeking->content = $request->content;
-        $seeking->seeking_thumbnail = $imagePath;
 
         // ファイルが送信された場合、DBに保存する
         if ($request->hasFile('seeking_thumbnail')) {
@@ -68,6 +65,8 @@ class SeekingController extends Controller
             $filename = time() . '.' . $seeking_thumbnail->getClientOriginalExtension();
             $seeking_thumbnail->storeAs('public/seeking_thumbnail', $filename); // ファイルを保存するディレクトリを指定する
             $seeking->seeking_thumbnail = $filename;
+        } else {
+          $seeking->seeking_thumbnail = 'default-thumbnail.png';
         }
 
 
@@ -143,7 +142,9 @@ class SeekingController extends Controller
         }
 
         // 募集の画像を削除
-        Storage::disk('public')->delete('seeking_thumbnail/' . $seeking->seeking_thumbnail);
+        if ($seeking->seeking_thumbnail !== 'default-thumbnail.png') {
+            Storage::disk('public')->delete('seeking_thumbnail/' . $seeking->seeking_thumbnail);
+        }
 
         // 募集を削除
         $seeking->delete();
